+++ src/main/java/com/ciphertool/genetics/util/FitnessEvaluator.java
@@ -3,5 +3,5 @@ package com.ciphertool.genetics.util;
 import com.ciphertool.genetics.entities.Chromosome;
 
 public interface FitnessEvaluator {
-	public int evaluate(Chromosome chromosome);
+	public Double evaluate(Chromosome chromosome);
 }
